backend/pyproject.toml
[project]
name = "deepfake-detector-backend"
version = "1.0.0"
description = "Backend API for deepfake image detection using Vision Transformer"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Deepfake Detector Team"}
]
keywords = ["deepfake", "detection", "machine-learning", "computer-vision", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "pillow>=10.3.0",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "transformers>=4.37.0",
    "accelerate>=0.26.0",
    "numpy>=1.26.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/deepfake-detector"
Documentation = "https://github.com/yourusername/deepfake-detector/docs"
Repository = "https://github.com/yourusername/deepfake-detector"
Issues = "https://github.com/yourusername/deepfake-detector/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312", "py313"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

# Made with Bob
